requirements.txt
aiogram==2.9
python-dotenv
asyncpg